style(structure): reorganize project directory layout
