+++ backend/internal/dto/announcement.go
