+++ backend/internal/global/global.go
