+++ backend/internal/models/announcement.go
