+++ backend/internal/router/announcement.go
@@ -1,13 +1,13 @@
 package router
 
 import (
-	"FashOJ_Backend/controllers"
+	"FashOJ_Backend/internal/controllers"
 
 	"github.com/gin-gonic/gin"
 )
 
 func setupAnnouncementRoutes(r *gin.RouterGroup) {
-	
+
 	announcement := r.Group("/announcement")
 	{
 		announcement.POST("", controllers.CreateAnnouncement)
