+++ backend/internal/router/auth.go
@@ -1,13 +1,13 @@
 package router
 
 import (
-	"FashOJ_Backend/controllers"
+	"FashOJ_Backend/internal/controllers"
 
 	"github.com/gin-gonic/gin"
 )
 
 func setupAuthRouter(r *gin.RouterGroup) {
-	
+
 	auth := r.Group("/auth")
 	{
 		auth.POST("/login", controllers.Login)
