+++ backend/internal/router/user.go
@@ -1,13 +1,13 @@
 package router
 
 import (
-	"FashOJ_Backend/controllers"
+	"FashOJ_Backend/internal/controllers"
 
 	"github.com/gin-gonic/gin"
 )
 
 func setupUserRoutes(r *gin.RouterGroup) {
-	
+
 	user := r.Group("/user")
 	{
 		user.POST("/changepermission", controllers.ChangePermission)
